cursor improvements (#1)
* rename for readability

* cursor improvements
