+++ .vscode/launch.json
@@ -10,7 +10,7 @@
             "request": "launch",
             "module": "uvicorn",
             "args": [
-                "app.main:app",
+                "src.main:app",
                 "--reload"
             ],
             "jinja": true
