+++ Dockerfile
@@ -13,9 +13,6 @@ COPY pyproject.toml poetry.lock ./
 
 RUN poetry install --without dev --no-root && rm -rf ${POETRY_CACHE_DIR}
 
-COPY . ./app
+COPY ./src ./src
 
-
-EXPOSE 80
-
-ENTRYPOINT ["poetry", "run", "uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "80"]
+ENTRYPOINT ["poetry", "run", "uvicorn", "src.main:app", "--host", "0.0.0.0", "--port", "80"]
