+++ src/parser.py
@@ -1,25 +1,7 @@
-import base64
-import json
 from email.message import Message
 from email.utils import parsedate_to_datetime
-from typing import Optional
 
 from .models import EmailMessageModel
-
-
-def encode_cursor(page: int, page_size: int) -> str:
-    cursor_data = {"page": page, "page_size": page_size}
-    cursor_str = json.dumps(cursor_data)
-    return base64.urlsafe_b64encode(cursor_str.encode()).decode()
-
-
-def decode_cursor(cursor: str) -> Optional[dict]:
-    try:
-        cursor_str = base64.urlsafe_b64decode(cursor.encode()).decode()
-        return json.loads(cursor_str)
-    except Exception as e:
-        print(f"Failed to decode cursor: {e}")
-        return None
 
 
 def parse_email_message(msg: Message) -> EmailMessageModel:
