+++ src/service.py
@@ -8,9 +8,9 @@ from datetime import datetime
 
 from .client import EmailClient
 from .imap_search_criteria import IMAPSearchCriteria
-from .parser import encode_cursor, decode_cursor, parse_email_message
+from .parser import parse_email_message
 from .models import (
-    Meta, PaginationMeta, PaginatedResponse, ApiResponse, EmailMessageModel, ImapServer
+    CursorModel, Meta, PaginationMeta, PaginatedResponse, ApiResponse, EmailMessageModel, ImapServer
 )
 
 ModelType = TypeVar('ModelType')
@@ -35,23 +35,13 @@ class EmailService:
         self,
         start_date: datetime,
         end_date: datetime,
-        page_size: int,
-        cursor: Optional[str] = None,
+        cursor: CursorModel,
         filter: Optional[Callable[[EmailMessageModel], bool]] = None
     ) -> ApiResponse[PaginatedResponse[EmailMessageModel]]:
         time_total = 0.0
 
         criteria = IMAPSearchCriteria().date_range(start_date, end_date)
         cache_key = self._generate_cache_key(criteria)
-
-        # 1. Interpret cursor
-        if cursor:
-            cursor_data = decode_cursor(cursor)
-            if cursor_data is None:
-                raise ValueError("Invalid cursor")
-            current_page = cursor_data['page']
-        else:
-            current_page = 1
 
         # 2. Get email ids
         email_ids = self.cache.get(cache_key)
@@ -78,9 +68,9 @@ class EmailService:
 
         # 3. Paginate emails
         total_items = len(email_ids)
-        offset = (current_page - 1) * page_size
+        offset = (cursor.page - 1) * cursor.page_size
         # Now that we have the entire list of emails, get only the segment of ids for the page requested
-        paginated_email_ids = email_ids[offset:offset + page_size]
+        paginated_email_ids = email_ids[offset:offset + cursor.page_size]
 
         with self._get_client() as client:
             emails, time_email = client.fetch_emails_by_ids(
@@ -100,21 +90,19 @@ class EmailService:
             self.logger.info(
                 f'Filtered out {page_total_items - filtered_total_items} emails')
             self.logger.info(f'Total images retrieved = {len()}')
-            if filtered_total_items == 0:
-                return ApiResponse(meta=Meta(status=HTTPStatus.NO_CONTENT, message='Filtered out all documents', request_time=time_total))
 
-        total_pages = (total_items + page_size - 1) // page_size
+        total_pages = (total_items + cursor.page_size - 1) // cursor.page_size
 
-        next_cursor = encode_cursor(
-            current_page + 1, page_size) if current_page < total_pages else None
-        prev_cursor = encode_cursor(
-            current_page - 1, page_size) if current_page > 1 else None
+        next_cursor = CursorModel(
+            page=cursor.page+1, page_size=cursor.page_size).encode()
+        prev_cursor = CursorModel(
+            page=cursor.page-1, page_size=cursor.page_size).encode() if cursor.page > 1 else None
 
         pagination_meta = PaginationMeta(
             total_items=total_items,
             total_pages=total_pages,
-            page_size=page_size,
-            current_page=current_page,
+            page_size=cursor.page_size,
+            current_page=cursor.page,
             next_cursor=next_cursor,
             prev_cursor=prev_cursor
         )
